[#3] refactor: improve asset directory structure
